Added basic tree code
Add basic tree code

Allows for traversal of character tree, adding and getting of words,
with associated scores.

Unsure of the best way to implement. A lot of the code is repeated
between add and get, but it seems like they'll diverge a bit. Julia
would probably be able to generate these functions with less duplication
than this, but not sure if it's possible in python. Currently going with
duplicating code in functions, but I may come back to this problem for
cleanup.